# Fix asset meta bottleneck