Fixes #6821, a performance issue introduced by #6769. 

In order to sync the original state to get the focal point, it would read the meta file. We would only try to read it if the file existed though, but on remote filesystems like S3 this would still result in an API request. The bottleneck is cause by all these new API requests.

This PR fixes the issue by checking whether the meta file is in the already locally cached file listing. It avoids those API requests.

Here are some benchmarks of the average times over 5 requests for a few simple pages I threw together.

**Asset browser:**
Before:  1.23s
After: 202ms

**Frontend page:**
Before: 1.09s
After: 222ms

**Publish form:**
Before: 2.6s
After: 257ms